Change dependencies

+++ kopen.c
@@ -243,9 +243,9 @@ static char **cmd2argv(const char *cmd)
 	for (i = beg + 1, argc = 0; i < end; ++i)
 		if (isspace(cmd[i]) && !isspace(cmd[i-1]))
 			++argc;
-	argv = (char**)calloc(argc + 2, sizeof(void*));
-	argv[0] = str = (char*)calloc(end - beg + 1, 1);
-	strncpy(argv[0], cmd + beg, end - beg);
+	argv = (char**)calloc((size_t)(argc + 2), sizeof(void*));
+	argv[0] = str = (char*)calloc((size_t)(end - beg + 1), 1);
+	strncpy(argv[0], cmd + beg, (size_t)(end - beg));
 	for (i = argc = 1; i < end - beg; ++i)
 		if (isspace(str[i])) str[i] = 0;
 		else if (str[i] && str[i-1] == 0) argv[argc++] = &str[i];
